+++ packages/site/src/examples/clientServer/server.ts
@@ -1,7 +1,12 @@
-import { ActionCall, applyAction, deserializeActionCall, getSnapshot } from "mobx-keystone"
+import {
+  applySerializedActionAndTrackNewModelIds,
+  getSnapshot,
+  SerializedActionCall,
+  SerializedActionCallWithModelIdOverrides,
+} from "mobx-keystone"
 import { createRootStore } from "../todoList/store"
 
-type MsgListener = (actionCall: ActionCall) => void
+type MsgListener = (actionCall: SerializedActionCallWithModelIdOverrides) => void
 
 class Server {
   private serverRootStore = createRootStore()
@@ -11,30 +16,34 @@ class Server {
     return getSnapshot(this.serverRootStore)
   }
 
-  onMessage(listener: (actionCall: ActionCall) => void) {
+  onMessage(listener: (actionCall: SerializedActionCallWithModelIdOverrides) => void) {
     this.msgListeners.push(listener)
   }
 
-  sendMessage(actionCall: ActionCall) {
+  sendMessage(actionCall: SerializedActionCall) {
     // the timeouts are just to simulate network delays
     setTimeout(() => {
       // apply the action over the server root store
       // sometimes applying actions might fail (for example on invalid operations
       // such as when one client asks to delete a model from an array and other asks to mutate it)
       // so we try / catch it
-      let applyActionSucceeded = false
+      let serializedActionCallToReplicate: SerializedActionCallWithModelIdOverrides | undefined
       try {
-        const deserializedActionCall = deserializeActionCall(actionCall, this.serverRootStore)
-        applyAction(this.serverRootStore, deserializedActionCall)
-        applyActionSucceeded = true
+        // we use this to apply the action on the server side and keep track of new model IDs being
+        // generated, so the clients will have the chance to keep those in sync
+        const applyActionResult = applySerializedActionAndTrackNewModelIds(
+          this.serverRootStore,
+          actionCall
+        )
+        serializedActionCallToReplicate = applyActionResult.serializedActionCall
       } catch (err) {
         console.error("error applying action to server:", err)
       }
 
-      if (applyActionSucceeded) {
+      if (serializedActionCallToReplicate) {
         setTimeout(() => {
-          // and distribute message
-          this.msgListeners.forEach(listener => listener(actionCall))
+          // and distribute message, which includes new model IDs to keep them in sync
+          this.msgListeners.forEach(listener => listener(serializedActionCallToReplicate!))
         }, 500)
       }
     }, 500)